Added moving plane

+++ src/plane.hpp
@@ -1,6 +1,6 @@
 #ifndef PLANE_H
 #define PLANE_H
 
-int define_plane(int z);
+int define_plane(float z);
 
 #endif
